requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
schedule>=1.2.0
ccxt>=4.0.0
pandas>=1.5.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
asyncio 